+++ packages/widgets/src/HeaderBar/Profile/ProfileHeader.js
@@ -1,10 +1,10 @@
 import { useConfig } from '@dhis2/app-runtime'
 import i18n from '@dhis2/d2-i18n'
 import propTypes from '@dhis2/prop-types'
+import { Box } from '@dhis2/ui-core'
 import React from 'react'
-import { ImageIcon } from '../ImageIcon.js'
 import { joinPath } from '../joinPath.js'
-import { TextIcon } from '../TextIcon.js'
+import { UserIcon } from '../UserIcon/UserIcon.js'
 
 const ProfileName = ({ children }) => (
     <div data-test="headerbar-profile-username">
@@ -91,10 +91,15 @@ ProfileDetails.propTypes = {
     name: propTypes.string,
 }
 
-export const ProfileHeader = ({ name, email, img }) => (
+export const ProfileHeader = ({ name, email, avatarUrl }) => (
     <div>
-        {img ? <ImageIcon src={img} /> : <TextIcon name={name} />}
-
+        <Box width="48px" height="48px">
+            <UserIcon
+                avatarUrl={avatarUrl}
+                name={name}
+                dataTest="headerbar-profile-menu-icon"
+            />
+        </Box>
         <ProfileDetails name={name} email={email} />
 
         <style jsx>{`
@@ -109,7 +114,7 @@ export const ProfileHeader = ({ name, email, img }) => (
 )
 
 ProfileHeader.propTypes = {
+    avatarUrl: propTypes.string,
     email: propTypes.string,
-    img: propTypes.string,
     name: propTypes.string,
 }